backend/requirements-full.txt
# Full deps for local dev (includes stocks). Use: pip install -r requirements-full.txt
fastapi
mangum
python-dotenv
uvicorn[standard]
google-genai
anthropic
openai
alpaca-py
pandas-ta
pytz
yfinance
